requirements.txt
# Nanakwaku Boateng Boakye-Akyeampong | Index: 10022200167
# CS4241 RAG project — manual pipeline (no LangChain / LlamaIndex)

streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
pypdf>=4.0.0
requests>=2.31.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
torch>=2.0.0
openai>=1.30.0
